split backend into multiple modules
