+++ jut-config.js
@@ -14,6 +14,5 @@
             answer: 42,
             foo: 'bar'
         }
-     }
-
- }
+    }
+}
